+++ Roman Bychkov/Lesson36/Lesson36.BL/Services/PersonsServices.cs
@@ -9,14 +9,26 @@ namespace Lesson36.BL.Services
         {
             _personsRepository = rep;
         }
-        public void Add(Person person)
+        public async Task Add(Person person)
         {
-            _personsRepository.Add(person);
+            if (string.IsNullOrWhiteSpace(person.FirstName))
+                throw new ArgumentException("Invalid person first name");
+            if (string.IsNullOrWhiteSpace(person.LastName))
+                throw new ArgumentException("Invalid person last name");
+            if (person.Age <= 16)
+                throw new ArgumentException("Invalid person age");
+            if(string.IsNullOrWhiteSpace(person.Gender))
+                throw new ArgumentException("Invalid person gender");
+            if (string.IsNullOrWhiteSpace(person.Address))
+                throw new ArgumentException("Invalid person adress");
+            await _personsRepository.Add(person);
         }
 
-        public void Delete(Person person)
+        public async Task Delete(Person person)
         {
-             _personsRepository.Delete(person);
+            if (person is null)
+                throw new ArgumentNullException("Person does not exist");
+            await _personsRepository.Delete(person);
         }
 
         public async Task<IEnumerable<Person>> Get()
@@ -29,9 +41,20 @@ namespace Lesson36.BL.Services
             return await _personsRepository.GetById(id);
         }
 
-        public void Update(Person person)
+        public async Task Update(Person person, int id)
         {
-            _personsRepository.Update(person);
+            var updatePerson = await GetById(id);
+            if (updatePerson is null)
+                throw new ArgumentNullException("Person does not exist");
+            else
+            {
+                updatePerson.FirstName = person.FirstName ?? updatePerson.FirstName;
+                updatePerson.LastName = person.LastName ?? updatePerson.LastName;
+                updatePerson.Age = person.Age < 16 ? updatePerson.Age : person.Age;
+                updatePerson.Gender = person.Gender ?? updatePerson.Gender;
+                updatePerson.Address = person.Address ?? updatePerson.Address;
+            }
+            await _personsRepository.Update(updatePerson);
         }
     }
 }
